requirements.txt
streamlit>=1.33.0
google-generativeai>=0.3.0
langchain-community>=0.0.34
langchain-core>=0.1.41
langchain>=0.1.14
python-dotenv>=1.0.0